# src/docs/openapi.dev.json
{
  "openapi": "3.0.3",
  "info": { "title": "Business API (Dev)", "version": "1.0.0-dev", "description": "Full specification including Dev/Debug endpoints" },
  "servers": [ { "url": "http://localhost:3000" } ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "apiKeyAuth": { "type": "apiKey", "in": "header", "name": "Authorization", "description": "Use 'Authorization: Token ck_...' for customer API key auth" }
    },
    "schemas": {
...
      "ApiKey": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "status": { "type": "string", "enum": ["active","revoked","suspended"] }, "createdAt": { "type": "string" } }, "required": ["id","name","status"] },
      "ApiKeyCreateResponse": { "type": "object", "properties": { "key": { "$ref": "#/components/schemas/ApiKey" }, "token": { "type": "string", "description": "Plaintext ck_ token (shown once)" } }, "required": ["token"] },
...
    }
  },
  "paths": {
...
    "/api/v1/customer/api-keys": { "get": { "tags": ["Customer"], "summary": "List API keys", "security": [{"bearerAuth":[]}], "responses": { "200": { "description": "Keys" } } }, "post": { "tags": ["Customer"], "summary": "Create API key", "security": [{"bearerAuth":[]}], "requestBody": { "required": false, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" } } } } } }, "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiKeyCreateResponse" } } } }, "401": { "description": "Unauthorized" } } } },
    "/api/v1/customer/api-keys/{id}": { "delete": { "tags": ["Customer"], "summary": "Revoke API key", "security": [{"bearerAuth":[]}], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ], "responses": { "200": { "description": "Revoked" }, "400": { "description": "Key not found" } } } },
...
  }
}